Refactoring of TaskServiceSession part 1. Moved some business logic into classes, ensure transactions are properly handled, added some new exceptions, catch exceptions in server, etc.
git-svn-id: http://anonsvn.jboss.org/repos/labs/labs/jbossrules/trunk@26014 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-process/drools-process-task/src/main/java/org/drools/task/event/TaskEvent.java
@@ -17,7 +17,7 @@ public abstract class TaskEvent extends EventObject implements Externalizable {
     
     public TaskEvent(long taskId) {
         super( taskId );
-        this.taskId = (Long) taskId;
+        this.taskId = taskId;
     }
     
     public void writeExternal(ObjectOutput out) throws IOException {
